Delete Indicator StudentResult and MeasLevel with fix

+++ src/main/java/pe/edu/pucp/sia/service/impl/MeasurementLevelServiceImpl.java
@@ -1,19 +1,25 @@
 package pe.edu.pucp.sia.service.impl;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import pe.edu.pucp.sia.model.Indicator;
 import pe.edu.pucp.sia.model.LevelDetail;
 import pe.edu.pucp.sia.model.Measurement;
 import pe.edu.pucp.sia.model.MeasurementLevel;
+import pe.edu.pucp.sia.model.MeasurementPlanLine;
 import pe.edu.pucp.sia.model.ResultsPerCard;
 import pe.edu.pucp.sia.model.Semester;
 import pe.edu.pucp.sia.model.Specialty;
+import pe.edu.pucp.sia.model.comparators.LevelDetailComparator;
+import pe.edu.pucp.sia.repository.IndicatorRepository;
 import pe.edu.pucp.sia.repository.LevelDetailRepository;
 import pe.edu.pucp.sia.repository.MeasurementLevelRepository;
+import pe.edu.pucp.sia.repository.MeasurementPlanLineRepository;
 import pe.edu.pucp.sia.repository.ResultsPerCardRepository;
 import pe.edu.pucp.sia.repository.SemesterRepository;
 import pe.edu.pucp.sia.requests.MPlanLineSpecialtySemesterRequest;
@@ -29,6 +35,12 @@ public class MeasurementLevelServiceImpl implements MeasurementLevelService {
 	
 	@Autowired
 	private ResultsPerCardRepository resultsPerCardRepository;
+	
+	@Autowired
+	private IndicatorRepository indicatorRepository;
+	
+	@Autowired
+    private MeasurementPlanLineRepository mplRepository;
 	
 	@Override
 	public ApiResponse listAll() {
@@ -59,6 +71,17 @@ public class MeasurementLevelServiceImpl implements MeasurementLevelService {
 		ApiResponse response = null;
 		try {
 			Integer id = measurementLevelRepository.save(ml).getId();
+			//Crea niveles de medición a cada indicador existente
+			Iterable<Indicator> list = indicatorRepository.findBystudentResultSpecialtyIdAndStudentResultSemesterIdOrderByCodeAsc(ml.getSpecialty().getId(),ml.getSemester().getId());
+			if (list!=null) {
+				ml.setId(id);
+				for (Indicator indicator: list) {
+					LevelDetail ld = new LevelDetail();
+					ld.setIndicator(indicator);
+					ld.setMeasurementLevel(ml);
+					levelDetailRepository.save(ld);
+				}
+			}
 			response = new ApiResponse(id,201);
 			
 		} catch(Exception ex) {
@@ -83,12 +106,22 @@ public class MeasurementLevelServiceImpl implements MeasurementLevelService {
 	public ApiResponse deleteMeasurementLevel(Integer id) {
 		ApiResponse response = null;
 		try {
-			List<LevelDetail> list = levelDetailRepository.findByMeasurementLevelId(id);
-			for (LevelDetail ld : list) {
-				levelDetailRepository.deleteLevelDetail(ld.getId());
-			}
-			measurementLevelRepository.deleteMeasurementLevel(id);
-			response = new ApiResponse(list,200);
+			//Verifica si hay planes de medición, que indica que el proceso ya empezo
+			MeasurementLevel ml = measurementLevelRepository.findById(id).get();
+			Iterator<MeasurementPlanLine> i = mplRepository.findByIndicatorStudentResultSpecialtyIdAndSemesterId(ml.getSpecialty().getId(),ml.getSemester().getId()).iterator();
+			if(!i.hasNext()) { //Si no hay
+									
+				List<LevelDetail> list = levelDetailRepository.findByMeasurementLevelId(id);
+				for (LevelDetail ld : list) {
+					levelDetailRepository.deleteLevelDetail(ld.getId());
+				}
+				measurementLevelRepository.deleteMeasurementLevel(id);
+				response = new ApiResponse(list,200);
+				
+			} else { //Si ya se empezó el proceso de medición
+				response = new ApiResponse(409,"No se puede eliminar niveles porque existen Planes de Medición");
+			}		
+			
 		} catch(Exception ex) {
 			response = new ApiResponse(500, ex.getMessage());
 		}
